Moved all cursor packages to their sibling wrapper packages
