+++ src/main/java/org/cufy/http/client/wrapper/ClientReqImpl.java
@@ -13,7 +13,7 @@
  *	See the License for the specific language governing permissions and
  *	limitations under the License.
  */
-package org.cufy.http.client.cursor;
+package org.cufy.http.client.wrapper;
 
 import org.cufy.http.Endpoint;
 import org.cufy.http.Request;
@@ -28,21 +28,21 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Objects;
 
 /**
- * An implementation of the interface {@link ClientRes}.
+ * An implementation of the interface {@link ClientReq}.
  *
  * @param <E> the type of the endpoint.
  * @author LSafer
  * @version 0.3.0
  * @since 0.3.0 ~2021.12.23
  */
-public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
+public class ClientReqImpl<E extends Endpoint> implements ClientReq<E> {
 	/**
-	 * The twin request cursor of this.
+	 * The twin response cursor of this.
 	 *
 	 * @since 0.3.0 ~2021.12.23
 	 */
 	@NotNull
-	protected final ClientReq<E> req;
+	protected final ClientRes<E> res;
 	/**
 	 * The current set client.
 	 *
@@ -79,15 +79,15 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 	@NotNull
 	protected Pipe<ClientRes<E>> pipe;
 	/**
-	 * The current set response.
+	 * The current set request.
 	 *
 	 * @since 0.3.0 ~2021.12.23
 	 */
 	@NotNull
-	protected Response response;
+	protected Request request;
 
 	/**
-	 * Construct a new client response wrapper with the most none code breaking defaults.
+	 * Construct a new client request wrapper with the most none code breaking defaults.
 	 * <br>
 	 * <br>
 	 * Note: the wrapper might not work if some necessary values are not set. This
@@ -98,7 +98,7 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 	 * @throws NullPointerException if the given {@code endpoint} is null.
 	 * @since 0.3.0 ~2021.12.23
 	 */
-	public ClientResImpl(
+	public ClientReqImpl(
 			@NotNull E endpoint
 	) {
 		Objects.requireNonNull(endpoint, "endpoint");
@@ -111,24 +111,24 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 				error.printStackTrace();
 		};
 		this.performer = null;
-		this.response = new Response();
-		this.req = new ClientReqDelegate();
+		this.request = new Request();
+		this.res = new ClientResDelegate();
 	}
 
 	/**
-	 * Construct a new client response wrapper with the given components.
+	 * Construct a new client request wrapper with the given components.
 	 *
 	 * @param endpoint the endpoint to be used.
-	 * @param response the initial response instance.
-	 * @param request  the initial request instance to be set (for {@link #req()}).
-	 * @throws NullPointerException if the given {@code endpoint} or {@code response} or
-	 *                              {@code request} is null.
+	 * @param request  the initial request instance to be set.
+	 * @param response the initial response instance (for {@link #res()}).
+	 * @throws NullPointerException if the given {@code endpoint} or {@code request} or
+	 *                              {@code response} is null.
 	 * @since 0.3.0 ~2021.12.23
 	 */
-	public ClientResImpl(
+	public ClientReqImpl(
 			@NotNull E endpoint,
-			@NotNull Response response,
-			@NotNull Request request
+			@NotNull Request request,
+			@NotNull Response response
 	) {
 		Objects.requireNonNull(endpoint, "endpoint");
 		Objects.requireNonNull(request, "request");
@@ -142,13 +142,13 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 				error.printStackTrace();
 		};
 		this.performer = null;
-		this.req = new ClientReqDelegate(request);
-		this.response = response;
+		this.request = request;
+		this.res = new ClientResDelegate(response);
 	}
 
 	@NotNull
 	@Override
-	public ClientRes<E> client(@Nullable Client client) {
+	public ClientReq<E> client(@Nullable Client client) {
 		Objects.requireNonNull(client, "client");
 		this.client = client;
 		return this;
@@ -162,7 +162,7 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 
 	@NotNull
 	@Override
-	public ClientRes<E> endpoint(@NotNull E endpoint) {
+	public ClientReq<E> endpoint(@NotNull E endpoint) {
 		Objects.requireNonNull(endpoint, "endpoint");
 		this.endpoint = endpoint;
 		return this;
@@ -176,7 +176,7 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 
 	@NotNull
 	@Override
-	public ClientRes<E> next(@NotNull Next<ClientRes<E>> next) {
+	public ClientReq<E> next(@NotNull Next<ClientRes<E>> next) {
 		Objects.requireNonNull(next, "next");
 		this.next = next;
 		return this;
@@ -190,7 +190,7 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 
 	@NotNull
 	@Override
-	public ClientRes<E> pipe(@NotNull Pipe<ClientRes<E>> pipe) {
+	public ClientReq<E> pipe(@NotNull Pipe<ClientRes<E>> pipe) {
 		Objects.requireNonNull(pipe, "pipe");
 		this.pipe = pipe;
 		return this;
@@ -204,27 +204,27 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 
 	@NotNull
 	@Override
-	public ClientReq<E> req() {
-		return this.req;
-	}
-
-	@NotNull
-	@Override
-	public ClientRes<E> response(@NotNull Response response) {
-		Objects.requireNonNull(response, "response");
-		this.response = response;
+	public ClientReq<E> request(@NotNull Request request) {
+		Objects.requireNonNull(request, "request");
+		this.request = request;
 		return this;
 	}
 
 	@NotNull
 	@Override
-	public Response response() {
-		return this.response;
+	public Request request() {
+		return this.request;
 	}
 
 	@NotNull
 	@Override
-	public ClientRes<E> strategy(@Nullable Strategy strategy) {
+	public ClientRes<E> res() {
+		return this.res;
+	}
+
+	@NotNull
+	@Override
+	public ClientReq<E> strategy(@Nullable Strategy strategy) {
 		this.performer = strategy;
 		return this;
 	}
@@ -236,131 +236,131 @@ public class ClientResImpl<E extends Endpoint> implements ClientRes<E> {
 	}
 
 	/**
-	 * A client request wrapper delegating to the hosting client response wrapper.
+	 * A client response wrapper delegating to the hosting client request wrapper.
 	 *
 	 * @author LSafer
 	 * @version 0.3.0
 	 * @since 0.3.0 ~2021.12.23
 	 */
-	public class ClientReqDelegate implements ClientReq<E> {
+	protected class ClientResDelegate implements ClientRes<E> {
 		/**
-		 * The current set request.
+		 * The current set response.
 		 *
 		 * @since 0.3.0 ~2021.12.23
 		 */
 		@NotNull
-		protected Request request;
+		protected Response response;
 
 		/**
-		 * Construct a new client request wrapper delegating to the variables in hosting
-		 * response wrapper.
+		 * Construct a new client response wrapper delegating to the variables in hosting
+		 * request wrapper.
 		 *
 		 * @since 0.3.0 ~2021.12.23
 		 */
-		protected ClientReqDelegate() {
-			this.request = new Request();
+		protected ClientResDelegate() {
+			this.response = new Response();
 		}
 
 		/**
-		 * Construct a new client request wrapper delegating to the variables in hosting
-		 * response wrapper.
+		 * Construct a new client response wrapper delegating to the variables in hosting
+		 * request wrapper.
 		 *
-		 * @param request the initial request instance to be set.
-		 * @throws NullPointerException if the given {@code request} is null.
+		 * @param response the initial response instance to be set.
+		 * @throws NullPointerException if the given {@code response} is null.
 		 * @since 0.3.0 ~2021.12.23
 		 */
-		protected ClientReqDelegate(@NotNull Request request) {
-			Objects.requireNonNull(request, "request");
-			this.request = request;
+		protected ClientResDelegate(@NotNull Response response) {
+			Objects.requireNonNull(response, "response");
+			this.response = response;
 		}
 
 		@NotNull
 		@Override
-		public ClientReq<E> client(@Nullable Client<ClientReq<E>, ClientRes<E>> client) {
+		public ClientRes<E> client(@Nullable Client<ClientReq<E>, ClientRes<E>> client) {
 			Objects.requireNonNull(client, "client");
-			ClientResImpl.this.client = client;
+			ClientReqImpl.this.client = client;
 			return this;
 		}
 
 		@NotNull
 		@Override
 		public Client<ClientReq<E>, ClientRes<E>> client() {
-			return ClientResImpl.this.client;
+			return ClientReqImpl.this.client;
 		}
 
 		@NotNull
 		@Override
-		public ClientReq<E> endpoint(@NotNull E endpoint) {
+		public ClientRes<E> endpoint(@NotNull E endpoint) {
 			Objects.requireNonNull(endpoint, "endpoint");
-			ClientResImpl.this.endpoint = endpoint;
+			ClientReqImpl.this.endpoint = endpoint;
 			return this;
 		}
 
 		@NotNull
 		@Override
 		public E endpoint() {
-			return ClientResImpl.this.endpoint;
+			return ClientReqImpl.this.endpoint;
 		}
 
 		@NotNull
 		@Override
-		public ClientReq<E> next(@NotNull Next<ClientRes<E>> next) {
+		public ClientRes<E> next(@NotNull Next<ClientRes<E>> next) {
 			Objects.requireNonNull(next, "next");
-			ClientResImpl.this.next = next;
+			ClientReqImpl.this.next = next;
 			return this;
 		}
 
 		@NotNull
 		@Override
 		public Next<ClientRes<E>> next() {
-			return ClientResImpl.this.next;
+			return ClientReqImpl.this.next;
 		}
 
 		@NotNull
 		@Override
-		public ClientReq<E> pipe(@NotNull Pipe<ClientRes<E>> pipe) {
+		public ClientRes<E> pipe(@NotNull Pipe<ClientRes<E>> pipe) {
 			Objects.requireNonNull(pipe, "pipe");
-			ClientResImpl.this.pipe = pipe;
+			ClientReqImpl.this.pipe = pipe;
 			return this;
 		}
 
 		@NotNull
 		@Override
 		public Pipe<ClientRes<E>> pipe() {
-			return ClientResImpl.this.pipe;
+			return ClientReqImpl.this.pipe;
 		}
 
 		@NotNull
 		@Override
-		public ClientReq<E> request(@NotNull Request request) {
-			Objects.requireNonNull(request, "request");
-			this.request = request;
+		public ClientReq<E> req() {
+			return ClientReqImpl.this;
+		}
+
+		@NotNull
+		@Override
+		public ClientRes<E> response(@NotNull Response response) {
+			Objects.requireNonNull(response, "response");
+			this.response = response;
 			return this;
 		}
 
 		@NotNull
 		@Override
-		public Request request() {
-			return this.request;
+		public Response response() {
+			return this.response;
 		}
 
 		@NotNull
 		@Override
-		public ClientRes<E> res() {
-			return ClientResImpl.this;
-		}
-
-		@NotNull
-		@Override
-		public ClientReq<E> strategy(@Nullable Strategy strategy) {
-			ClientResImpl.this.performer = strategy;
+		public ClientRes<E> strategy(@Nullable Strategy strategy) {
+			ClientReqImpl.this.performer = strategy;
 			return this;
 		}
 
 		@NotNull
 		@Override
 		public Strategy strategy() {
-			return ClientResImpl.this.performer;
+			return ClientReqImpl.this.performer;
 		}
 	}
 }
