--- src/main/java/org/cufy/http/concurrent/cursor/package-info.java
@@ -1,23 +1,0 @@
-/*
- *	Copyright 2021 Cufy and ProgSpaceSA
- *
- *	Licensed under the Apache License, Version 2.0 (the "License");
- *	you may not use this file except in compliance with the License.
- *	You may obtain a copy of the License at
- *
- *	    http://www.apache.org/licenses/LICENSE-2.0
- *
- *	Unless required by applicable law or agreed to in writing, software
- *	distributed under the License is distributed on an "AS IS" BASIS,
- *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *	See the License for the specific language governing permissions and
- *	limitations under the License.
- */
-/**
- * A sub-package of {@link org.cufy.http.cursor} for concurrent models.
- *
- * @author LSafer
- * @version 0.3.0
- * @since 0.3.0 ~2021.12.23
- */
-package org.cufy.http.concurrent.cursor;
