+++ src/test/kotlin/org/cufy/http/endpoint/EndpointTest2.kt
@@ -4,12 +4,12 @@ import org.cufy.http.Endpoint
 import org.cufy.http.body.json
 import org.cufy.http.client.Http
 import org.cufy.http.concurrent.Strategy
-import org.cufy.http.cursor.Req
-import org.cufy.http.cursor.component1
-import org.cufy.http.cursor.component2
+import org.cufy.http.wrapper.Req
 import org.cufy.http.endpoint.MyEndpoint.name
 import org.cufy.http.json.JsonString
 import org.cufy.http.okhttp.OkEngine
+import org.cufy.http.wrapper.component1
+import org.cufy.http.wrapper.component2
 import org.cufy.http.wrapper.endpoint
 
 object MyEndpoint : Endpoint {
